[INFRA]实现了后端部分功能

+++ backend/CarSharing/src/main/java/com/tongji/carsharing/Entity/Address.java
@@ -13,8 +13,8 @@ import java.sql.Timestamp;
 @Builder
 public class Address {
 
-    private Long addressId;          // 地址ID
-    private Long addressUserId;      // 用户ID (外键，关联到 user 表)
+    private Integer addressId;          // 地址ID
+    private Integer addressUserId;      // 用户ID (外键，关联到 user 表)
     private String addressDetail;    // 详细地址
     private Timestamp addressCreateTime; // 创建时间
 }
